+++ src/main/kotlin/calculator/ExpressionParser.kt
@@ -1,4 +1,4 @@
-package step2
+package calculator
 
 class ExpressionParser(private val expression: List<String>) {
