+++ src/main/kotlin/carGame/Configuration.kt
@@ -1,0 +1,10 @@
+package carGame
+
+import carGame.domain.strategy.MoveStrategy
+import carGame.domain.strategy.RandomMoveStrategy
+
+object Configuration {
+    fun moveStrategy(): MoveStrategy {
+        return RandomMoveStrategy
+    }
+}
