+++ src/main/kotlin/carGame/domain/CarGame.kt
@@ -1,6 +1,6 @@
-package step3.domain
+package carGame.domain
 
-import step3.domain.strategy.MoveStrategy
+import carGame.domain.strategy.MoveStrategy
 
 class CarGame(private val moveStrategy: MoveStrategy) {
     fun play(carCount: Int, tryCount: Int): GameResult {
