+++ src/test/kotlin/study/PersonTest.kt
@@ -1,4 +1,4 @@
-package step1
+package study
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.assertAll
